refactor(wa-sqlite): move types and use index.ts for import
