+++ examples/playground/src/playground/hooks/usePlaygroundSync.ts
@@ -22,7 +22,7 @@ import {
 } from '@treecrdt/sync-protocol/browser';
 import { treecrdtSyncV0ProtobufCodec } from '@treecrdt/sync-protocol/protobuf';
 import { wrapDuplexTransportWithCodec, type DuplexTransport } from '@treecrdt/sync-protocol/transport';
-import type { TreecrdtClient } from '@treecrdt/wa-sqlite/client';
+import type { TreecrdtClient } from '@treecrdt/wa-sqlite';
 
 import { hexToBytes16, type AuthGrantMessageV1 } from '../../sync-v0';
 import {
