+++ packages/treecrdt-wa-sqlite/package.json
@@ -14,6 +14,10 @@
       "import": "./dist/client.js",
       "types": "./dist/client.d.ts"
     },
+    "./types": {
+      "import": "./dist/types.js",
+      "types": "./dist/types.d.ts"
+    },
     "./opfs": {
       "import": "./dist/opfs.js",
       "types": "./dist/opfs.d.ts"
