+++ packages/treecrdt-wa-sqlite/src/index.ts
@@ -4,7 +4,34 @@ import type { MaterializationEvent } from '@treecrdt/interface/engine';
 import { dbGetText } from './sql.js';
 import type { Database } from './types.js';
 
-export type { Database } from './types.js';
+export type {
+  ClientMode,
+  ClientOptions,
+  Database,
+  RuntimeMode,
+  StorageMode,
+  TreecrdtAssets,
+  TreecrdtClient,
+  TreecrdtClientAuthApi,
+  TreecrdtRuntime,
+  TreecrdtStorage,
+} from './types.js';
+
+export type {
+  OpfsSupport,
+  OpfsVfsKind,
+  OpfsVfsOptions,
+  OpenOptions,
+} from './opfs.js';
+export {
+  clearOpfsStorage,
+  createOpfsVfs,
+  detectOpfsSupport,
+  openWithStorage,
+  opfsStorageExists,
+} from './opfs.js';
+
+export { CLIENT_CLOSED_ERROR, createTreecrdtClient } from './client.js';
 
 function createRunner(db: Database): SqliteRunner {
   return {
