+++ packages/treecrdt-wa-sqlite/src/open.ts
@@ -1,6 +1,6 @@
 import { createOpfsVfs, type OpfsVfsKind } from './opfs.js';
 import { createWaSqliteApi } from './index.js';
-import type { Database } from './index.js';
+import type { Database } from './types.js';
 import { makeDbAdapter } from './db.js';
 import type { TreecrdtAdapter } from '@treecrdt/interface';
 import type { MaterializationEvent } from '@treecrdt/interface/engine';
